Move common ctl functions to the src/ directory
