+++ tiering/Makefile.mk
@@ -4,7 +4,9 @@
 lib_LTLIBRARIES += tiering/libmemtier.la \
                    # end
 
-tiering_libmemtier_la_SOURCES = tiering/ctl.c \
+tiering_libmemtier_la_SOURCES = \
+                  memkind/internal/ctl.h \
+                  tiering/ctl.c \
                   tiering/ctl.h \
                   tiering/memtier.c \
                   tiering/memtier_log.c \
